requirements.txt
groq
streamlit
python-dotenv
requests
beautifulsoup4
langchain
langchain-groq
langchain-core 
langchain-community 
langchain-openai
chromadb
openai 
langchain-google-genai
langchain-anthropic
# for the finance search agent
unstructured 
libmagic 
python-magic 
python-magic-bin
langchain-text-splitters
# faiss 
faiss-cpu
sentence-transformers
unstructured[pdf]